+++ dtslint/ts3.5/ReadonlyRecord.ts
@@ -180,3 +180,10 @@ _.isSubrecord(eqNumber)(recordString) // $ExpectType (me: Readonly<Record<string
 
 _.lookup('a', recordString) // $ExpectType Option<number>
 _.lookup('a') // $ExpectType <A>(r: Readonly<Record<string, A>>) => Option<A>
+
+//
+// elem
+//
+
+_.elem(eqNumber)(1, recordString) // $ExpectType boolean
+_.elem(eqNumber)(1) // $ExpectType (fa: Readonly<Record<string, number>>) => boolean