+++ test/Record.ts
@@ -282,6 +282,9 @@ describe('Record', () => {
   it('elem', () => {
     assert.deepStrictEqual(_.elem(eqNumber)(1, { a: 1, b: 2 }), true)
     assert.deepStrictEqual(_.elem(eqNumber)(3, { a: 1, b: 2 }), false)
+
+    assert.deepStrictEqual(_.elem(eqNumber)(1)({ a: 1, b: 2 }), true)
+    assert.deepStrictEqual(_.elem(eqNumber)(3)({ a: 1, b: 2 }), false)
   })
 
   it('fromFoldableMap', () => {