+++ CHANGELOG.md
@@ -26,6 +26,8 @@ high state of flux, you're at risk of it changing without notice.
     - make `union` data-last (@gcanti)
     - make `zip` data-last (@gcanti)
     - make `cons` data-last (@gcanti)
+  - `Map`
+    - make `elem` data-last (@gcanti)
   - `ReadonlyArray`
     - export `unfold` (@gcanti)
     - make `lookup` data-last (@gcanti)
@@ -35,6 +37,8 @@ high state of flux, you're at risk of it changing without notice.
     - make `union` data-last (@gcanti)
     - make `zip` data-last (@gcanti)
     - make `cons` data-last (@gcanti)
+  - `ReadonlyMap`
+    - make `elem` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)