+++ CHANGELOG.md
@@ -54,9 +54,11 @@ high state of flux, you're at risk of it changing without notice.
   - `ReadonlyRecord`
     - make `isSubrecord` data-last (@gcanti)
     - make `lookup` data-last (@gcanti)
+    - make `elem` data-last (@gcanti)
   - `Record`
     - make `isSubrecord` data-last (@gcanti)
     - make `lookup` data-last (@gcanti)
+    - make `elem` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)