+++ CHANGELOG.md
@@ -17,6 +17,14 @@ high state of flux, you're at risk of it changing without notice.
 # 2.6.5
 
 - **Polish**
+  - `Array`
+    - export `unfold` (@gcanti)
+  - `ReadonlyArray`
+    - export `unfold` (@gcanti)
+
+# 2.6.5
+
+- **Polish**
   - export a pipeable `wither` function from all modules which admit a `Witherable` instance (@gcanti)
   - export a pipeable `wilt` function from all modules which admit a `Witherable` instance (@gcanti)